feat: Put checking into a crate referenced from the binary
